requirements.txt
django
django-allauth
djangorestframework
social-auth-app-django
groq
python-dotenv
